tests/requirements-test.txt
# profit_scout_pipeline/tests/requirements-test.txt
# Combined requirements for unit and integration tests

# Testing framework & utilities
pytest
pytest-asyncio 
pytest-mock     
pytest-cov      
pytest-flask   

# Google Cloud stack
google-cloud-storage
google-cloud-secret-manager
google-generativeai
google-cloud-bigquery  
google-cloud-core   
google-auth          

# Core application / ML / Finance stack dependencies
Flask          
numpy           
tenacity        
pandas         
yfinance       
websockets      
tqdm            